requirements/requirements.txt
asgiref==3.2.3
Babel==2.7.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.2
django-cors-headers==3.2.0
django-environ==0.4.5
django-localflavor==2.2
django-phonenumber-field==3.0.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.11.0
future==0.18.2
idna==2.8
itypes==1.1.0
jedi==0.15.2
Jinja2==2.10.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
parso==0.5.2
phonenumbers==8.11.1
phonenumberslite==8.10.14
pluggy==0.13.1
python-language-server==0.19.0
python-stdnum==1.12
pytz==2019.3
requests==2.22.0
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.7
